Add files via upload
Here is a mock expense tracker that I decided to do. There's some back end stuff and integration stuff that can be used. Will import to actual expense tracker when I have time. NOTE: the node modules folders are not in here. This is because there were too many files in that folder to be uploaded at once